consolidating root src files
- moving library + App into a single `viewer` file with both
