+++ webpack.config.js
@@ -10,7 +10,7 @@ const fileName = `${packageName}.min.js`;
 const banner = `${libraryName} - ${packageName} - ${VERSION} \nprovided and maintained by ${AUTHOR} \nLICENSE MIT`;
 
 module.exports = {
-  entry: "./src/library.js",
+  entry: "./src/viewer.js",
   output: {
     filename: fileName,
     library: libraryName,
